requirements.txt
gradio<=4.40
python-dotenv
bs4
jinja2
markdownify

# for LLM
openai
huggingface_hub

langchain
langchain_community
langchain_openai
langchain_huggingface

# for RAG
# sentence-transformers # depends on torch, which is large
pypdf
ruamel.yaml
chromadb
langchain_text_splitters
langchain_chroma

# for Azure
azure-cognitiveservices-speech

# for agent
tool2schema